-enh rename vl6180x registers according to google style guide
