+++ .gitignore
@@ -31,3 +31,5 @@
 *.out
 *.app
 build/
+.idea/
+cmake-build-debug/
