+++ modules/hal_sensors/sensor_compression/src/vl6180x_registers.h
